@@ -25,24 +25,23 @@
   },
   "homepage": "https://github.com/kubevious/helpers#readme",
   "dependencies": {
-    "dotenv": "^8.2.0",
-    "mysql2": "^2.2.5",
-    "the-lodash": "^2.0.9",
-    "the-logger": "^2.0.9",
-    "the-promise": "^2.0.13"
+    "dotenv": "^16.1.3",
+    "mysql2": "^3.3.3",
+    "the-lodash": "^2.1.2",
+    "the-logger": "^2.1.1",
+    "the-promise": "^2.1.3"
   },
   "devDependencies": {
-    "@types/mocha": "^8.0.3",
-    "@types/mysql2": "github:types/mysql2",
-    "@types/node": "^14.6.0",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "^20.2.5",
     "@types/should": "^13.0.0",
-    "@typescript-eslint/eslint-plugin": "^3.10.0",
-    "@typescript-eslint/parser": "^3.10.0",
-    "eslint": "^7.7.0",
-    "mocha": "^8.1.1",
-    "prettier": "^2.1.0",
+    "@typescript-eslint/eslint-plugin": "^5.59.8",
+    "@typescript-eslint/parser": "^5.59.8",
+    "eslint": "^8.41.0",
+    "mocha": "^10.2.0",
+    "prettier": "^2.8.8",
     "should": "^13.2.3",
-    "ts-node": "^9.0.0",
-    "typescript": "^4.0.2"
+    "ts-node": "^10.9.1",
+    "typescript": "^5.1.3"
   }
 }
