@@ -7,11 +7,11 @@ export function calculateObjectHash(obj : any) : Buffer
         throw new Error('NO Object');
     }
 
-    var str = _.stableStringify(obj);
+    const str = _.stableStringify(obj);
 
     const sha256 = crypto.createHash('sha256');
     sha256.update(str);
-    var value = sha256.digest();
+    const value = sha256.digest();
     return value;
 }
 
